dashboard/app/requirements.txt
dash==2.3.1
pandas==1.4.2
plotly==5.7.0
gunicorn==20.1.0
Flask==2.1.2
google-cloud-storage==2.3.0
gcsfs==2022.3.0